HW-5 - Refactoring Project

+++ class13/Sangmeebee/MyProject/app/src/main/java/com/example/myproject/data/model/Items.kt
@@ -19,5 +19,5 @@ data class Items(
     @SerializedName("actor")
     val actor: String,
     @SerializedName("userRating")
-    val userRating: Double
+    val userRating: Float
 )
